# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Wrenn API
  description: AI agent execution platform API.
  version: "0.2.0"
servers:
  - url: http://localhost:8080
    description: Local development
security: []
paths:
...
  /v1/teams:
    get:
      summary: List teams for the authenticated user
      operationId: listTeams
      tags: [teams]
      security:
        - sessionAuth: []
      responses:
        "200":
          description: Teams the user belongs to, each with their role
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamWithRole"

    post:
      summary: Create a new team
      operationId: createTeam
      tags: [teams]
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  description: 1-128 chars; A-Z a-z 0-9 space _
      responses:
        "201":
          description: Team created (caller is owner)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamWithRole"
        "400":
          description: Invalid team name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/teams/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Team ID (must match the JWT's team_id)

    get:
      summary: Get team info and member list
      operationId: getTeam
      tags: [teams]
      security:
        - sessionAuth: []
      responses:
        "200":
          description: Team details with members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamDetail"
        "403":
          description: JWT team does not match requested team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Team not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      summary: Rename the team
      operationId: renameTeam
      tags: [teams]
      security:
        - sessionAuth: []
      description: Admin or owner role required (verified from DB).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "204":
          description: Renamed
        "400":
          description: Invalid team name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete the team
      operationId: deleteTeam
      tags: [teams]
      security:
        - sessionAuth: []
      description: |
        Owner only. Soft-deletes the team and destroys all running/paused/starting
        capsules. All DB records are preserved. The team slug is permanently reserved.
      responses:
        "204":
          description: Team deleted
        "403":
          description: Caller is not the owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/teams/{id}/members/{uid}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: uid
        in: path
        required: true
        schema:
          type: string
        description: Target user ID

    patch:
      summary: Update member role
      operationId: updateMemberRole
      tags: [teams]
      security:
        - sessionAuth: []
      description: |
        Admin or owner required. Valid target roles: admin, member.
        The owner's role cannot be changed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [admin, member]
      responses:
        "204":
          description: Role updated
        "403":
          description: Insufficient role or attempt to modify owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User is not a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Remove a member
      operationId: removeTeamMember
      tags: [teams]
      security:
        - sessionAuth: []
      description: Admin or owner required. Owner cannot be removed.
      responses:
        "204":
          description: Member removed
        "403":
          description: Insufficient role or attempt to remove owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User is not a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/teams/{id}/leave:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string

    post:
      summary: Leave the team
      operationId: leaveTeam
      tags: [teams]
      security:
        - sessionAuth: []
      description: The owner cannot leave; they must delete the team instead.
      responses:
        "204":
          description: Left the team
        "403":
          description: Owner cannot leave
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/admin/teams:
    get:
      summary: List all teams (admin)
      operationId: adminListTeams
      tags: [admin]
      security:
        - sessionAuth: []
      responses:
        "200":
          description: Teams list
          content:
            application/json:
              schema:
                type: array
                items: {type: object}
...
  /v1/admin/teams/{id}:
    delete:
      summary: Delete a team (admin)
      operationId: adminDeleteTeam
      tags: [admin]
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
      responses:
        "204":
          description: Deleted
...
components:
...
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for capsule lifecycle operations. Create via POST /v1/api-keys.

    sessionAuth:
      type: apiKey
      in: cookie
      name: wrenn_sid
      description: |
        Opaque session cookie set by POST /v1/auth/login, /v1/auth/activate, or
        the OAuth callback. HttpOnly, Secure, SameSite=Strict. Idle window 6h,
        absolute lifetime 24h. State-changing requests also require an
        X-CSRF-Token header matching the wrenn_csrf cookie (double-submit).
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Double-submit CSRF token whose value must match the wrenn_csrf cookie.
        Required on all non-GET requests authenticated via session cookie.
        Not required for API key auth.

    hostTokenAuth:
      type: apiKey
      in: header
      name: X-Host-Token
      description: Host JWT returned from POST /v1/hosts/register or POST /v1/hosts/auth/refresh. Valid for 7 days.
  schemas:
...
    Team:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
          description: Immutable 12-char hex slug (e.g. a1b2c3-d1e2f3)
        created_at:
          type: string
          format: date-time
    TeamWithRole:
      allOf:
        - $ref: "#/components/schemas/Team"
        - type: object
          properties:
            role:
              type: string
              enum: [owner, admin, member]
    TeamMember:
      type: object
      properties:
        user_id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [owner, admin, member]
        joined_at:
          type: string
          format: date-time
    TeamDetail:
      type: object
      properties:
        team:
          $ref: "#/components/schemas/Team"
        members:
          type: array
          items:
            $ref: "#/components/schemas/TeamMember"
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...